More tests for MovementController

+++ client/src/config.h
@@ -4,6 +4,6 @@
 static const int MAX_UPDATE_RATE = 100;
 static constexpr int MAX_FPS = 60 - 1;
 
-static const float MOVEMENT_DELAY_MILLIS = 400;
+static const int MOVEMENT_DELAY_MILLIS = 400;
 
 #endif // CONFIG
